Dependency Injection by Reference Object
There are 2 classes those are "A" and "B". Suppose class B is contain some other functionalities and class A contain some different functionalities. But those functionality are provide a different features from each other those are important.
But when we create the object of class "B" in Class "A" it means Class "A" is dependent on Class "B" and class "B" is a Dependency class. To configure we need a XML file which is refer the Class "A" and "B". And in the main class we can provide a XML path of configuring file, then we can achieve the our expected output.